Concerns #32 - Moved the chart report component into its own folder
