+++ src/app/components/reports/chart-report/chart-report-configuration.component.spec.ts
