+++ src/app/components/simulation/simulation.component.ts
@@ -1,5 +1,4 @@
 import { Component, Input, ContentChild } from '@angular/core';
-import { ChartReportComponent } from '../reports/chart-report.component';
 import { AccountsComponent } from '../accounts/accounts.component';
 import { QuotesService } from 'src/app/services/quotes/quotes.service';
 import { Simulation } from 'src/app/model/core/simulation';
